디테일 완성
