+++ src/components/Tag.tsx
@@ -3,7 +3,7 @@ import Link from 'next/link'
 
 const Tag = ({ text }: { text: string }) => {
   return (
-    <Link href={`/tags/${text}`}>
+    <Link href={`/tags/${text}/pages/1`}>
       <a className="mr-3 text-sm font-medium text-blue-500 uppercase hover:text-blue-600 dark:hover:text-blue-400">
         {text.split(' ').join('-')}
       </a>
