+++ src/components/layouts/List.tsx
@@ -45,7 +45,7 @@ export default function ListLayout({
                     <div>
                       <h3 className="text-2xl font-bold leading-8 tracking-tight">
                         <CustomLink
-                          href={`/blog/${slug}`}
+                          href={`/${slug}`}
                           className="text-gray-900 dark:text-gray-100"
                         >
                           {title}
