+++ src/utils/Markdown.ts
@@ -1,8 +1,4 @@
-import { join } from 'path'
-
 import renderToString from 'next-mdx-remote/render-to-string'
-import { statSync, readdirSync, readFile } from 'promise-fs'
-import frontMatter from 'front-matter'
 import remarkMath from 'remark-math'
 // eslint-disable-next-line @typescript-eslint/ban-ts-comment
 // @ts-ignore
@@ -14,17 +10,10 @@ import toc from 'remark-toc'
 // eslint-disable-next-line @typescript-eslint/ban-ts-comment
 // @ts-ignore
 import slug from 'remark-slug'
-import memoize from 'memoizee'
 import { MdxRemote } from 'next-mdx-remote/types'
 import visit from 'unist-util-visit'
 
-import { FrontMatter, Post, TagWithCount } from '#commons/types'
-import MDXComponents from '#components/MDXComponents'
-
-const DIR_REPLACE_STRING = '/posts'
-const POST_PATH = `${process.cwd()}${DIR_REPLACE_STRING}`
-
-export const getAllPosts = memoize(retreiveAllPosts)
+import MDXComponents from '../components/MDXComponents'
 
 type TokenType =
   | 'tag'
@@ -51,28 +40,6 @@ const tokenClassNames: { [key in TokenType]: string } = {
   function: 'text-code-blue',
   boolean: 'text-code-red',
   comment: 'text-gray-400 italic',
-}
-
-function getFilesRecursively(path: string) {
-  const getFiles = (path: string) =>
-    readdirSync(path)
-      .map((name) => join(path, name))
-      .filter((path: string) => statSync(path).isFile())
-
-  const isDirectory = (path: string) => statSync(path).isDirectory()
-
-  const getDirectories = (path: string) =>
-    readdirSync(path)
-      .map((name) => join(path, name))
-      .filter(isDirectory)
-
-  const dirs = getDirectories(path)
-
-  const files: string[] = dirs
-    .map((dir) => getFilesRecursively(dir))
-    .reduce((a, b) => a.concat(b), [])
-
-  return files.concat(getFiles(path)).filter((f) => f.endsWith('.md'))
 }
 
 // const imgToJSX = (_: any) => (tree: any) => {
@@ -105,82 +72,6 @@ function getFilesRecursively(path: string) {
 //     },
 //   )
 // }
-
-export async function getAllDraftPosts() {
-  const files = getFilesRecursively(POST_PATH).reverse()
-  const draftPosts: Array<string> = []
-
-  for await (const f of files) {
-    const file = await readFile(f, { encoding: 'utf8' })
-    const { attributes } = frontMatter(file)
-    const fm: FrontMatter = attributes as any
-    const { published } = fm
-
-    const slug = f
-      .slice(f.indexOf(DIR_REPLACE_STRING) + DIR_REPLACE_STRING.length + 1)
-      .replace('.md', '')
-
-    if (!published) {
-      draftPosts.push(slug)
-    }
-  }
-
-  console.table(draftPosts)
-  return draftPosts
-}
-
-export async function retreiveAllPosts(): Promise<Array<Post>> {
-  const files = getFilesRecursively(POST_PATH).reverse()
-  const posts: Array<Post> = []
-
-  for await (const f of files) {
-    const file = await readFile(f, { encoding: 'utf8' })
-    const { attributes, body } = frontMatter(file)
-    const fm: FrontMatter = attributes as any
-    const { tags: fmTags, published, date } = fm
-
-    const slug = f
-      .slice(f.indexOf(DIR_REPLACE_STRING) + DIR_REPLACE_STRING.length + 1)
-      .replace('.md', '')
-
-    if (published) {
-      const tags = (fmTags || []).map((tag) => tag.trim())
-
-      const result: Post = {
-        frontmatter: {
-          ...fm,
-          tags,
-          date: new Date(date).getTime() - 9 * 60 * 60 * 1000, // 한국시간
-        },
-        body,
-        fields: {
-          slug,
-        },
-        path: f,
-      }
-
-      posts.push(result)
-    }
-  }
-
-  return posts.sort((a, b) => b.frontmatter.date - a.frontmatter.date)
-}
-
-export async function getAllTagsFromPosts(): Promise<Array<TagWithCount>> {
-  const tags = (await getAllPosts()).reduce((prev, curr) => {
-    curr.frontmatter.tags.forEach((tag) => {
-      prev.push(tag)
-    })
-    return prev
-  }, [] as string[])
-
-  const tagWithCount = [...new Set(tags)].map((tag) => ({
-    tag,
-    count: tags.filter((t) => t === tag).length,
-  }))
-
-  return tagWithCount.sort((a, b) => b.count - a.count)
-}
 
 export async function parseMarkdownToMDX(
   body: string,
